(1) SDF(Soooa Data Format) 추가 	- Datum, SDF, DataReader, DataInputLayer, IO 추가 	- SDF로부터 데이터를 처리할 수 있는 입력 레이어 추가 	- Mnist, 일반 이미지 데이터셋에서 SDF 생성할 수 있는 예제 추가 (2) Data에 set_host_data, offset 적용 (3) Accuracy Layer 추가

+++ SoooA/src/frcnn/frcnn_common.h
@@ -363,7 +363,7 @@ template <typename Dtype>
 static void np_array_value_by_index_array(const std::vector<std::vector<Dtype>>& array,
 		const uint32_t axis, const std::vector<uint32_t>& inAxisIndex,
 		std::vector<Dtype>& result) {
-	SASSERT0(axis >= 0 && axis < 2);
+	SASSERT0(axis >= uint32_t(0) && axis < uint32_t(2));
 
 	const uint32_t numArrayElem = array.size();
 	SASSERT0(numArrayElem > 0);
@@ -542,7 +542,7 @@ static void np_maximum(const Dtype value, const std::vector<Dtype>& array,
 template <typename Dtype>
 static void np_argmax(const std::vector<std::vector<Dtype>>& array, const uint32_t axis,
 		std::vector<uint32_t>& result) {
-	SASSERT0(axis >= 0 && axis < 2);
+	SASSERT0(axis >= uint32_t(0) && axis < uint32_t(2));
 
 	const uint32_t numArrayElem = array.size();
 	SASSERT0(numArrayElem >= 1);
